+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added the pagination to listUsers query
+
 ## [1.18.1] - 2022-04-28
 
 ### Fixed